1.21.4 -> 1.21.5

+++ src/main/java/eu/ha3/presencefootsteps/sound/generator/TerrestrialStepSoundGenerator.java
@@ -36,7 +36,7 @@ class TerrestrialStepSoundGenerator implements StepSoundGenerator {
     // Airborne
     protected boolean isAirborne;
 
-    protected float lastFallDistance;
+    protected double lastFallDistance;
     protected float lastReference;
     protected boolean isImmobile;
     protected long timeImmobile;
